Fix some bugs.

+++ gsv_tts/GPT_SoVITS/GPT/utils.py
@@ -15,7 +15,7 @@ def logits_to_probs(
     pre_tokens_lens: Optional[torch.Tensor] = None,
     temperature: float = 1.0,
     top_k: Optional[int] = None,
-    top_p: Optional[int] = None,
+    top_p: Optional[float] = None,
     repetition_penalty: float = 1.0,
 ):
     if previous_tokens is not None and repetition_penalty != 1.0:
